third commit

+++ contracts/SupplyChain.sol
@@ -11,8 +11,8 @@ contract SupplyChain {
 
     struct Item {
         string name;
-        uint256 sku;
-        uint256 price;
+        uint sku;
+        uint price;
         State state;
         address payable seller;
         address payable buyer;
